squashme: rename imports
