+++ bases/renku_data_services/data_api/main.py
@@ -6,7 +6,7 @@ from sanic import Sanic
 from sanic.log import logger
 from sanic.worker.loader import AppLoader
 
-from renku_data_services.config import Config
+from renku_data_services.app_config import Config
 from renku_data_services.data_api.app import register_all_handlers
 from renku_data_services.migrations.core import run_migrations_for_app
 from renku_data_services.storage.rclone import RCloneValidator
