+++ components/renku_data_services/app_config/__init__.py
@@ -1,2 +1,2 @@
 """Configuration parsing utilities."""
-from renku_data_services.config.config import Config, DBConfig  # noqa: F401, F403
+from renku_data_services.app_config.config import Config  # noqa: F401, F403
