+++ components/renku_data_services/db_config/__init__.py
@@ -1,0 +1,2 @@
+"""Configuration parsing utilities."""
+from renku_data_services.db_config.config import DBConfig  # noqa: F401, F403
