+++ pyproject.toml
@@ -9,7 +9,8 @@ packages = [
   { include = "renku_data_services/data_api", from = "bases" },
   { include = "renku_data_services/keycloak_sync", from = "bases" },
   { include = "renku_data_services/authn", from = "components" },
-  { include = "renku_data_services/config", from = "components" },
+  { include = "renku_data_services/db_config", from = "components" },
+  { include = "renku_data_services/app_config", from = "components" },
   { include = "renku_data_services/authz", from = "components" },
   { include = "renku_data_services/authz_migrations", from = "components" },
   { include = "renku_data_services/crc", from = "components" },
