+++ test/bases/renku_data_services/data_api/test_storage.py
@@ -6,7 +6,7 @@ from sanic import Sanic
 from sanic_testing.testing import SanicASGITestClient
 
 from renku_data_services.authn.dummy import DummyAuthenticator
-from renku_data_services.config import Config
+from renku_data_services.app_config import Config
 from renku_data_services.data_api.app import register_all_handlers
 from renku_data_services.storage.rclone import RCloneValidator
