+++ test/components/renku_data_services/data_api/test_config.py
@@ -1,6 +1,6 @@
 from unittest.mock import MagicMock
 
-import renku_data_services.config.config as conf
+import renku_data_services.app_config.config as conf
 from renku_data_services.authn.dummy import DummyAuthenticator
 from renku_data_services.k8s.clients import DummyCoreClient, DummySchedulingClient
