+++ test/components/renku_data_services/db/test_sqlalchemy_pool_repo.py
@@ -17,7 +17,7 @@ from pydantic import ValidationError
 
 import renku_data_services.base_models as base_models
 from renku_data_services import errors
-from renku_data_services.config import Config
+from renku_data_services.app_config import Config
 from renku_data_services.crc import models
