+++ test/components/renku_data_services/db/test_sqlalchemy_user_repo.py
@@ -12,7 +12,7 @@ import pytest
 from hypothesis import HealthCheck, given, settings
 
 import renku_data_services.base_models as base_models
-from renku_data_services.config import Config
+from renku_data_services.app_config import Config
 from renku_data_services.crc import models
